Move TuyaQuirkBuilder out of MCU, use AttrDefs
