+++ zhaquirks/tuya/__init__.py
@@ -975,6 +975,16 @@ class TuyaPowerConfigurationCluster3AA(TuyaPowerConfigurationCluster):
     }
 
 
+class TuyaPowerConfigurationCluster4AA(TuyaPowerConfigurationCluster):
+    """PowerConfiguration cluster for devices with 4 AA."""
+
+    _CONSTANT_ATTRIBUTES = {
+        PowerConfiguration.AttributeDefs.battery_size.id: 3,
+        PowerConfiguration.AttributeDefs.battery_rated_voltage.id: 15,
+        PowerConfiguration.AttributeDefs.battery_quantity.id: 4,
+    }
+
+
 class TuyaThermostat(CustomDevice):
     """Generic Tuya thermostat device."""
